pyproject.toml
[tool.poetry]
name = "doc-llm-qa"
version = "0.1.0"
description = ""
authors = ["juicyjung <travelandi01@gmail.com>"]
readme = "README.md"
packages = [{include = "doc_llm_qa"}]

[tool.poetry.dependencies]
python = "3.10"
streamlit = "^1.28.0"
langchain = "^0.0.325"
pypdf = "^3.17.0"
python-dotenv = "^1.0.0"
libhwp = "^0.2.0"
openai = "^1.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
